+++ .github/CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Improvements
 
+* Apply controlled-adjoint and adjoint-controlled operators with native C++ implementation
+  [(#1063)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1063)
+
 * Capture execution via `dev.eval_jaxpr` can now be used with `jax.jit` and `jax.vmap`.
   [(#1055)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1055)